# src/backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  models.CarsPic:
    properties:
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      fileName:
        type: string
      id:
        type: integer
      updatedAt:
        type: string
    type: object
...
  models.OwnerCarsInfo:
    properties:
      brand:
        type: string
      carPics:
        items:
          $ref: '#/definitions/models.CarsPic'
        type: array
      cartype:
        type: string
      model:
        type: string
      useLogs:
        items:
          $ref: '#/definitions/models.UserCar'
        type: array
      year:
        type: integer
    type: object
...
  models.UserCar:
    properties:
      UserId:
        type: integer
      carId:
        type: integer
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      endTime:
        type: string
      id:
        type: integer
      reason:
        description: |-
          0: Request     1:Pending for payment     2:Paid but not used
          3.During using 4.After using             5.Finished
        type: string
      startTime:
        type: string
      status:
        type: integer
      totalPrice:
        type: integer
      updatedAt:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: cheayuki13@gmail.com
  description: This is a car rental system API.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Car Rental System API
  version: "1.0"
paths:
...
  /api/v1/service/info/bookId/{id}:
    get:
      consumes:
      - application/json
      description: Get booking information by bookId
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserCar'
        "400":
          description: Invalid car ID
          schema:
            type: string
        "404":
          description: UserCar not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get booking info by bookId
      tags:
      - Service
  /api/v1/service/info/ownerId/{id}:
    get:
      consumes:
      - application/json
      description: Get booking information by ownerId
      parameters:
      - description: Owner ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.OwnerCarsInfo'
            type: array
        "400":
          description: Invalid Owner ID
          schema:
            type: string
        "500":
          description: Failed To Find Your Cars!
          schema:
            type: string
      summary: Get booking info by ownerId
      tags:
      - Service
  /api/v1/service/info/userId/{id}:
    get:
      consumes:
      - application/json
      description: Get booking information by userId
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UserCar'
            type: array
        "400":
          description: Invalid User ID
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Get booking info by userId
      tags:
      - Service
  /api/v1/service/status/approve/{id}:
    post:
      consumes:
      - application/json
      description: Approve a booking request by car owner
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Request approved successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Not A Pending Request!
          schema:
            type: string
        "404":
          description: UserCar not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Approve a booking request
      tags:
      - Service
  /api/v1/service/status/decline/{id}:
    post:
      consumes:
      - application/json
      description: Decline a booking request by car owner
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Request declined and deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Not A Pending Request!
          schema:
            type: string
        "404":
          description: UserCar not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Decline a booking request
      tags:
      - Service
  /api/v1/service/user/bookCar:
    post:
      consumes:
      - application/json
      description: Book a new car for rental
      parameters:
      - description: Rent request details
        in: body
        name: rentRequest
        required: true
        schema:
          $ref: '#/definitions/models.UserCar'
      produces:
      - application/json
      responses:
        "200":
          description: Car rented successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Car is not available
          schema:
            type: string
        "404":
          description: Car not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Book a new car
      tags:
      - Service
  /api/v1/service/user/returnCar:
    post:
      consumes:
      - application/json
      description: Return a rented car
      parameters:
      - description: Return request details
        in: body
        name: returnRequest
        required: true
        schema:
          $ref: '#/definitions/models.UserCar'
      produces:
      - application/json
      responses:
        "200":
          description: Car rented successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Relation between User and Car not found
          schema:
            type: string
        "404":
          description: Car not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Return a car
      tags:
      - Service
  /api/v1/service/user/status/decline/{id}:
    post:
      consumes:
      - application/json
      description: User decline a booking request
      parameters:
      - description: Booking ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Request declined and deleted successfully
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid Book ID
          schema:
            type: string
        "404":
          description: UserCar not found
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: User decline a booking request
      tags:
      - Service
...
swagger: "2.0"